api/pyproject.toml
[tool.pyright]
pythonVersion = "3.11.3"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
#Help generate the rest of the file
name = "solaceai-prototype"
version = "0.8.0"
readme = "README.md"
license = {text = 'Apache-2.0'}
description = "Python package to embed the Solace-ai (Ai2 Scholar QA) functionality in another application"
authors = [{ name = "Amanpreet Singh", email = "amanpreets@allenai.org" },
           { name = "Sergey Feldman", email = "sergey@allenai.org"},
            { name = "Joseph Chang", email = "josephc@allenai.org"},
            { name = "Jesus Dominguez", email = "jesus.dominguez@kcl.ac.uk"}]

dependencies = [
    "fastapi==0.115.8",
    "gunicorn==21.2.0",
    "httpx",
    "pydantic==2.*",
    "python-json-logger==2.0.4",
    "uvicorn[standard]",
    "nora_lib==1.0.0",
    "modal==1.0.2",
    "pandas",
    "jsonlines",
    "googlesearch-python",
    "backoff",
    "numpy",
    "pandas",
    "requests",
    "openai==1.106.1",
    "bs4",
    "fuzzy-match",
    "googlesearch-python",
    "tqdm",
    "google-cloud-storage==2.18.2",
    "filelock==3.16.1",
    "litellm==1.76.2",
    "pandas==2.2.2",
    "diskcache==5.6.3",
    "langsmith==0.1.142",
    "anyascii==0.3.2",
    "tenacity==9.1.2"
]
[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "wheel",
]
all = [
    "torch==2.3.1",
    "transformers==4.44.1",
    "sentencepiece",
    "hf-transfer==0.1.6",
    "huggingface_hub==0.23.4",
    "sentence-transformers==3.0.1",
    "peft",
    "black",
    "mypy",
    "wheel",
]
[project.urls]
Homepage = "https://scholarqa.allen.ai/"
Documentation = "https://github.com/allenai/ai2-scholarqa-lib"
Repository = "https://github.com/allenai/ai2-scholarqa-lib"

[tool.setuptools.packages.find]
exclude = ["internal"]
include = ["solaceai*"]

[tool.black]
preview = true

[tool.mypy]
strict = true
ignore_missing_imports = true

